+++ NEWS.md
@@ -16,6 +16,23 @@
 
 * Columns specified as having type "number" (requested via `col_number()` or `"number"` or `'n'`) or "skip" (requested via `col_skip()` or `"skip"` or `_` or `-`) now work in the case where 0 rows of data are parsed (#427, #540, #548).
 
+* `vroom()`, `vroom_lines()`, and `vroom_fwf()` now close and destroy (instead
+  of leak) the connection in the case where opening the connection fails due to,
+  e.g., a nonexistent URL (#488).
+
+* vroom takes the recommended approach for phasing out usage of the non-API
+  entry points `SETLENGTH` and `SET_TRUELENGTH` (#582).
+
+* If there is insufficient space for the tempfile used when reading from a
+  connection (affects delimited and fixed width parsing, from compressed files
+  and URLs), that is now reported as an error and no longer segfaults (#544).
+
+* `vroom(..., n_max = 0, col_names = c(...))` with a connection (compressed
+  file, URL, raw connection) no longer produces a "negative length vectors are
+  not allowed" error or crashes R (#539).
+
+* `vroom_fwf(..., n_max = 0)` with a connection no longer segfaults (#590).
+
 # vroom 1.6.7
 
 * `locale(encoding =)` now warns, instead of errors, when the encoding cannot be found in `iconvlist()` return value. This removes an unnecessary blocker on platforms like Alpine Linux where the output doesn't reflect actual capabilities.